main: move percentage filter to mirrors
Move the percentage filter to the only user of it and add a test for
basic use cases.
